Fix part of bug #470
openMSX-debugger crashes when the OPL3Cartridge2 extensions are used
  https://sourceforge.net/p/openmsx/bugs/470/

When connecting the debugger to openMSX, the connection dialog shows a list of
running openMSX instances (when there's more than one instance). To get a clue
of which instance to pick, the machine name and the output of the 'guess_title'
script is shown in the connect dialog.

In case the OPL3Cartridge2 is used, the 'guess_title' script generated an error
(see other part of this fix for more details on the script). And the crash was
caused because the debugger had a bug in the error-handling code-path in the
connection dialog.

This patch fixes part of the problem: the debugger no longer crashes when an
error occurs in some of the commands that are exchanged between debugger and
openmsx in the connection-phase. The 'guess_title' script itself will be fixed
in the main openmsx repo.

I *think* this patch fixes the problem. At least valgrind no longer complains
about reads from already deleted objects. It would still be nice if someone
with more Qt experience can review this fix.