Replace asprintf() with arg()
In 6615490f sprintf() was replaced by asprintf() to address deprecation warnings,
however the latter seemed to cause issues on macOS [1] for unknown reasons.

Since even for asprintf() the documentation warns against using it in new code,
I am replacing it with arg(), which is by the way also what most of the rest of the
debugger code uses.

[1] https://www.msx.org/forum/msx-talk/openmsx/please-help-testing-upcoming-openmsx-release?page=46#comment-394809