Fix track pad scrolling
On my Mac’s track pad, scrolling the disassembly and hex views does not work well.
This is because the wheel delta is divided but the remainder is discarded. Due to this,
they do not scroll when moving one’s fingers across the track pad too slowly.

This remembers the remainder and accumulates it with the delta.