Fix parsing of HiTech C MAP files.
I found some CP/M Hi-Tech C map files and these couldn't be parsed anymore.
After debugging it appeared that it didn't find the "Symbol Table" line. So I
loosened the check on that line a bit, which made it work.
As for the cross compiler's map files: they have more output in the file
preceeding the "magic" line. So I adjusted the code to skip any lines before
that magic line.

Fixes #52.