Remove invalid address from creation of condition
Creation of conditions shouldn't send -1 as an invalid address back to the emulator. If it does, the result is this wrong condition:

cond#1 65535 { condition_text }

Which is not what we want. The condition is a number and the condition text becomes the command (which will be ignored by the debugger). If we fixed createSetCommand by sending no address at all instead of an invalid one, it would set { debug break } as the default command:

cond#1 { condition_text } { debug break }