BreakpointViewer doesn't reload breakpoints anymore
BreakpointViewer now stores references to Breakpoints (Breakpoint, Watchpoints and Conditions) through 3 maps of new data structure "BreakpointData" because just storing references in a hidden column in QTableWidget is not flexible or powerful enough. The new data structure allows "BreakpointViewer" to sync with the emulator only when reconnected and not every time a new breakpoint is created or modified, which is more efficient. Also, since the table of breakpoints is not erased every time, items will not move around without the user's consent (they will not be destroyed and recreated, but reused).

minor changes:
* renamed processXXXX functions to parseXXXX because it is less of a generic name;
* hidden 6th table column stores breakpoint name (id) instead of index;
* more generic and useful setTextField function instead of item->setText();
* message box displayed when error received from openMSX;
* debugger now waits for openMSX response before performing changes;
* to avoid rows moving around, editing or adding a breakpoint now automatically turns sorting off.