Refactor parseXXX() functions in BreakpointViewer
Replace
  bool return-value + output-parameter
with
  std::optional<> return-val;ue

In pseudo code:
Before:
    bool parseValue(..., int& result);
    ...
    int result;
    if (parseValue(..., result)) {
        use(result);
    }
After:
    std::optional<int> parseValue(...);
    ...
    if (auto result = parseValue(...)) {
        use(*result);
    }

The main advantage is that it's no longer needed to define a (typically
uninitialized) 'result' variable upfront. It's now also more difficult
to forget to check the status (ok/failure) and then accidentally use the
uninitialized variable. As a bonus, often, the scope of the 'result'
variable can be smaller.

--

This patch also contains a bugfix:
Before there was an expression like:
    value < 0 && value > 3
But this is always false. Clang gives a warning on this:
    overlapping comparisons always evaluate to false [-Wtautological-overlap-compare]
but gcc does not.

Looking at the context of this expression, I think it should have been:
    !(0 <= value && value <= 3)