Added 'ranges.h' to simplify usage of STL algorithms
Our 'ranges.h' header contains a partial and limited
'reimplementation'(*) of the c++20 <ranges> header. Only a very small
amount of algorithms are present, only those that we actually use, more
can be added when needed. Also not the full functionality is
reimplemented, only the subset that can easily be obtained using c++17
(but e.g. nothing realted to 'concepts' which is a new c++20 language
feature).

(*) 'Reimplementation' may not be the best word. It's more like a
different API to the same underlying algorithm. And indeed our
implementations mostly only adapt some parameters and then delegate to
the corresponding std++17 algorithm.

In the future, when we switch to c++20, we can:
* replace   #include "ranges.h"   with   #include <ranges>
* and       ranges::xxx(...);     with   std::ranges::xxx(...);
And then remove our 'ranges.h' header.

(For now) the main advantage of these new algorithms is ease of use. For
example:

before:
    std::fill(std::begin(myArray), std::end(myArray), 123);
after:
    ranges::fill(myArray, 123);

before:
    std::lower_bound(myVector.begin(), myVector.end(), value,
        [](const MyStruct& s, uint32_t value) {
            return s.value < value;
        });
after:
    ranges::lower_bound(myVector, value, {}, &MyStruct::value);