Added QScrollArea to SwitchingWidget
Also tried some fancier popup menu on the SplitterHandle, but this
causes a lot of crashes. I assume that the handler is already free'ed while the
mouseReleaseEvent is still being processed causing these SegFaults.

Also I got divide by zero errors when creating new StackViewer with an open
openMSX connection, it seems to handle the command reply before the widget is
fully initialized???

Also the changing of background (when connecting/disconnecting the emulator) in
the scrollarea isn't eassily triggered. If you resize by hand it will correctly
change the color not covered by the single widget, but simply setting an new
background doesn't cause an update, neither doesn explicitely calling update()
or repaint().