Introduced a QuickGuide explaining the splitter
This widget simply contains a richtext label to explain the new workings
of the debugger.

The idea is that this is the default widget for a new SwitchingWidget
until the user is familiar with the workings and we have a setting to
switch this behaviour to a more regular blender like splitting of the
current area.