Json save and load of widget in SwitchingWidget.
This way we can 'correctly' perfom a split so that both the old and new
widgets have the same options selected. We can also save some of these
settings when we save layouts.

Unfortunately for stuff like the DebuggableViewer the actual text of the
combobox depends on the debuggables that the selected openMSX presents
to the debugger, so saving the selection in a layout will not work.
(Unless we connect to an openMSX that presents the same debuggables
before we even try to restore the layout? That would be a nice
improvement...)