Fixed missing QMetaObject::connectSlotsByName
When the 'slots:' indicator is removed in the header files the names do
no longer end up in the QMetaObject. This is not needed with the new Qt5
way of hooking up signal/slots, but the ui files use this to autoconnect
the slots to GUI elements by the on_<widgetname>_<signalname> naming
convention.
This no longer worked so we are fixing that here.
The VDPCommandRegViewer will be fixed more correctly later :-)